Refactor Loader into Loading Component
